+++ quetzal/app/api/data/file.py
@@ -14,7 +14,9 @@ from quetzal.app.helpers.google_api import get_object, get_data_bucket
 from quetzal.app.helpers.files import split_check_path, get_readable_info
 from quetzal.app.helpers.pagination import paginate
 from quetzal.app.api.exceptions import APIException
-from quetzal.app.models import BaseMetadataKeys, Family, Workspace, Metadata
+from quetzal.app.models import (
+    BaseMetadataKeys, Family, FileState, Workspace, Metadata
+)
 from quetzal.app.security import (
     PublicReadPermission, ReadWorkspacePermission, WriteWorkspacePermission
 )
@@ -61,6 +63,13 @@ def create(*, wid, content=None, user, token_info=None):
                                   'should not happen and will be reported to '
                                   'the administrator')
 
+    # Manage temporary parameter
+    temporary = (request.args.get('temporary', 'false').capitalize() == 'True')
+    if temporary:
+        state = FileState.TEMPORARY
+    else:
+        state = FileState.READY
+
     #  Create metadata object
     meta = Metadata(id_file=uuid4(), family=base_family)
     md5, size = get_readable_info(content)
@@ -73,6 +82,7 @@ def create(*, wid, content=None, user, token_info=None):
         'checksum': md5,
         'date': _now(),
         'url': '',
+        'state': state.name,
     }
     db.session.add(meta)
 
@@ -84,6 +94,62 @@ def create(*, wid, content=None, user, token_info=None):
     db.session.commit()
 
     return meta.json, codes.created
+
+
+def delete(*, wid, uuid, user, token_info=None):
+    workspace = Workspace.get_or_404(wid)
+
+    if not WriteWorkspacePermission(wid).can():
+        raise APIException(status=codes.forbidden,
+                           title='Forbidden',
+                           detail='You are not authorized to modify metadata on this workspace')
+
+    if not workspace.can_change_metadata:
+        # See note on 412 code and werkzeug on top of workspace.py file
+        raise APIException(status=codes.precondition_failed,
+                           title=f'Cannot update metadata of file',
+                           detail=f'Cannot update metadata of files to a '
+                                  f'workspace on {workspace.state.name} state')
+
+    # Before marking the file for deletion, we must ensure that this workspace
+    # has all the metadata families that reference this file
+    latest_meta_committed = Metadata.get_latest_global(uuid)
+    related_families = set(m.family.name for m in latest_meta_committed)
+    if related_families > set(f.name for f in workspace.families):
+        # Note the > is the superset operator
+        raise APIException(status=codes.precondition_failed,
+                           title='Cannot delete file',
+                           detail='Cannot delete file in this workspace '
+                                  'because the workspace does not use all the '
+                                  'metadata families associated with this file. '
+                                  'These families are: ' +
+                                  ', '.join(related_families) + '.')
+
+    # For the other families, deleting means erasing all key:value entries except id
+    for family in workspace.families:
+        if family.name not in related_families:
+            continue
+
+        latest = Metadata.get_latest(uuid, family)
+        if family.name == 'base':
+            # For the base family, deleting means changing its state
+            latest.update({'state' : FileState.DELETED.name})
+            db.session.add(latest)
+        else:
+            # For other families, it means deleting all keys except id
+            if latest.family.workspace is None:
+                # The metadata is global, not from this workspace
+                latest = latest.copy()
+                latest.json = {'id': uuid}
+                latest.family = family
+                db.session.add(latest)
+            else:
+                # The metadata is from this workspace
+                latest.json = {'id': uuid}
+                db.session.add(latest)
+
+    db.session.commit()
+    return None, codes.accepted
 
 
 def update_metadata(*, wid, uuid, body):