+++ quetzal/app/models.py
@@ -26,17 +26,6 @@ roles_users_table = db.Table('roles_users',
                              UniqueConstraint('fk_user_id', 'fk_role_id'))
 
 
-@enum.unique
-class BaseMetadataKeys(enum.Enum):
-    ID = 'id'
-    FILENAME = 'filename'
-    PATH = 'path'
-    SIZE = 'size'
-    CHECKSUM = 'checksum'
-    DATE = 'date'
-    URL = 'url'
-
-
 class Role(db.Model):
     id = db.Column(db.Integer(), primary_key=True)
     name = db.Column(db.String(80), unique=True)
@@ -102,6 +91,26 @@ class User(UserMixin, db.Model):
         return f'<User {self.username}>'
 
 
+@enum.unique
+class FileState(enum.Enum):
+    READY = 'ready'
+    TEMPORARY = 'temporary'
+    DELETED = 'deleted'
+
+
+@enum.unique
+class BaseMetadataKeys(enum.Enum):
+    ID = 'id'
+    FILENAME = 'filename'
+    PATH = 'path'
+    SIZE = 'size'
+    CHECKSUM = 'checksum'
+    DATE = 'date'
+    URL = 'url'
+    STATE = 'state'
+
+
+@enum.unique
 class WorkspaceState(enum.Enum):
     INITIALIZING = 'initializing'
     READY = 'ready'